// back/src/enums/event.ts
export enum EventName {
  UserCreate = "user:create",
  UserRetrieve = "user:retrieve",
  UserDelete = "user:delete",
  BoardUpdate = "board:update",
  RoomCreate = "room:create",
  RoomJoin = "room:join",
  RoomMove = "room:move",
  RoomInvite = "room:invite",
  RoomInviteAnswer = "room:invite:answer",
  RoomReady = "room:ready",
  RoomStart = "room:start",
  RoomFinish = "room:finish",
  RoomDestroy = "room:destroy",
}